refactor(leaderboard): move exportCSV to utility module and finalize test coverage
