+++ src/utils/csv.js
@@ -77,3 +77,45 @@ export function exportCSV(devs, headers = []) {
   a.remove();
   setTimeout(() => URL.revokeObjectURL(url), 100);
 }
+
+
+/**
+ * Exports data to a secure CSV file.
+ * @param {Array<Object>} devs - The data source array.
+ */
+function exportCSV(devs) {
+  // Ensure global columns fallback or fail gracefully
+  const headers =
+    typeof CSV_EXPORT_COLUMNS !== "undefined" ? CSV_EXPORT_COLUMNS : [];
+
+  // FIX: Removed the devs.length === 0 check to restore original behavior.
+  // Allowing empty arrays ensures users still download a clean header-only CSV when no matches are found.
+  if (!Array.isArray(devs) || headers.length === 0) return;
+
+  // Build rows dynamically and safely
+  const rows = devs.map((d) =>
+    headers
+      .map((h) => {
+        const cellValue = d ? d[h] : "";
+        return csvCell(cellValue);
+      })
+      .join(","),
+  );
+
+  // Combine headers and rows with a UTF-8 BOM (\uFEFF) for Excel compatibility
+  const csv = ["\uFEFF" + headers.map(csvCell).join(","), ...rows].join("\r\n");
+  const blob = new Blob([csv], { type: "text/csv;charset=utf-8;" });
+  const url = URL.createObjectURL(blob);
+
+  // Trigger secure client-side download
+  const a = document.createElement("a");
+  a.href = url;
+  a.download = `rankistan-leaderboard-${new Date().toISOString().slice(0, 10)}.csv`;
+  a.style.display = "none";
+  document.body.appendChild(a);
+  a.click();
+
+  // Instant DOM cleanup
+  a.remove();
+  setTimeout(() => URL.revokeObjectURL(url), 100);
+}
